requirements.txt
# Core Flask dependencies
Flask==3.1.0
flask-cors==5.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
itsdangerous==2.2.0

# ASGI Server
asgiref==3.8.1
uvicorn==0.34.0
h11==0.14.0

# Utilities
click==8.1.8
colorama==0.4.6
blinker==1.9.0

# Rich logging
rich==13.7.0

# Report generation
reportlab==4.0.9
matplotlib==3.8.3

# Additional recommended packages
python-dotenv==1.0.0

# Excel processing
pandas==2.2.0
openpyxl==3.1.2
xlrd==2.0.1

# Supabase database
supabase==2.22.0
